Add support for turbolinks
This change does two main things:
- element references are recached after next-page results are loaded
- any requests initiated before a turbolinks pageload are ignored if
they complete after the results are loaded

The element reference to the container element is made by attaching a
unique data attribute to it: “data-jquery-infinite-pages-container” and
setting the value to a unique id for the instance of InfinitePages.

Requests are ignored by keeping a queue of timestamps marking when they
begin and keeping a timestamp of the last time turbolinks changed the
page, invalidating all before that last page-load.